lawgatapi/requirements.txt
asgiref==3.8.1
boto3==1.39.2
botocore==1.39.2
Django==5.2.4
django-storages==1.14.6
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
jmespath==1.0.1
mysqlclient==2.2.7
pillow==11.3.0
PyJWT==2.9.0
python-dateutil==2.9.0.post0
python-decouple==3.8
s3transfer==0.13.0
six==1.17.0
sqlparse==0.5.3
tzdata==2025.2
urllib3==2.5.0
gunicorn
whitenoise